Create README.md
# BEST CLASSIFIER?
##ABSTRACT
This report explores two inquiries. To start with, for a given choice of datasets, would we be able to say what is the best classifier in terms of good predictions? What amount does the appropriate response rely upon the choice of datasets? How much does the answer depend on our computational constraints? We examine this inquiry utilizing datasets from the UCI repository. Second, are there some other potential routes through which we can improve the performance, considering the computational efficiency as well? Also, after having a brief idea of some of the machine learning concepts, have you tried your knowledge to some other datasets?

##INTRODUCTION
Supervised learning, one of the categories of Machine learning, is further classified as a Classification or Regression problem depending upon the dependent variable: categorical or discrete. Before diving into building models, it is necessary to know answers for rising questions: what our data looks like, are there any missing values, does it require scaling/normalization, what are the types of features, and are all required in building models and lot more. Our very first step is to load the data sets in a format easy to interpret. We used pandas for reading and manipulation of data since it provides hands-on features to load data in tabular format (data frame) and it has built-in methods to summarize the data set. We even checked for class imbalance by printing the frequency of each class for each data set, required for classification. This helps in choosing appropriate metrics for evaluation.

We made different functions for reading, preprocessing by applying appropriate techniques and splitting the data, and process all the files the same way maintaining consistency throughout the project.

Now that we have clean the data, we can feed it to defined classifiers. We can train on the training data and evaluate the model on the testing data. But, this will lead to overfitting as we are tuning the hyperparameters based on the testing data. Instead, a better approach is to use a left out training data called validation data and evaluate the model on it. Finally, predict the model on test data. This way we can avoid overfitting the model. We have used hyperparameters selection techniques like GridSearchCV and RandomSearchCV.

Once we have the best parameters for each model, finally we can fit the model on the best hyperparameters and check for accuracy. Here, we deal with different metrics based on the application.

##METHODOLOGY & EXPERIMENTAL RESULTS
This section deals with the common strategy applied to both classification problems. **Data loading**: We proposed a common method for classification tasks to read the whole dataset. **Reprocessing**: In this step, we clean the data by handling missing values. There are many techniques to do so like using the mean value, most frequent value or any value very small or high can be used to fill missing values. Also, there can be columns with categorical features and are required to be in a numeric format for most of the Machine learning, though so algorithms like decision trees can handle them. So, we have used label encoding to convert to numeric values. **Splitting**: In this method, we split the dataset into training and testing.